+++ src/main/scala/com/leysoft/products/domain/Product.scala
@@ -1,3 +1,3 @@
 package com.leysoft.products.domain
 
-case class Product(id: Long, name: String, stock: Double)
+case class Product(id: Int, name: String, stock: Double)
